requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==5.2.12
django-crispy-forms==2.6
gunicorn==25.2.0
pillow==12.1.1
psycopg2-binary==2.9.11
python-dotenv==1.2.2
requests==2.33.0
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.12.0
cloudinary==1.42.2
django-cloudinary-storage==0.3.0
python-telegram-bot==21.10
discord.py==2.4.0
